Module Vcs.Err

Payload of the exception raised by Vcs functions.

type t

The type of errors raised by Vcs.

Under the hood, it is lazily constructed human-readable information which also carry some context (a sort of a high level stack trace manipulated by the programmers of Vcs).

It is not meant to be matched on, but rather to be printed on stderr (or perhaps logged).

val sexp_of_t : t -> Base.Sexp.t

sexp_of_t t forces the lazy message, and allow printing the information contained by t.

val to_string_hum : t -> Base.string

to_string_hum t is a convenience wrapper around t |> sexp_of_t |> Sexp.to_string_hum.

val create_s : Base.Sexp.t -> t

val to_error : t -> Base.Error.t

Inject t into Base.Error.t. This is useful if you'd like to use Vcs inside the Or_error monad.

val add_context : t -> step:Base.Sexp.t -> t

Add a step of context into the stack trace contained by the error.

val init : Base.Error.t -> step:Base.Sexp.t -> t

This is useful if you are starting from an Error.t initially.