Module Vcs.Log

A log is a complete listing of the structure of the dag.

It contains all the commits, along with the parents of each commit.

module Line : sig ... end

type t = Line.t Base.list

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val sexp_of_t : t -> Sexplib0.Sexp.t

Accessors

val roots : t -> Rev.t Base.list

Return the list of all init commits.