Module Vcs.Name_status

module Change : sig ... end

type t = Change.t Base.list

val sexp_of_t : t -> Sexplib0.Sexp.t

val files : t -> Base.Set.M(Vcs__.Path_in_repo).t

Returns the set of files that are involved by the changes, either at src or dst.

Returns the set of files that are involved by the changes, that are also present at src or dst respectively.

val files_at_src : t -> Base.Set.M(Vcs__.Path_in_repo).t

val files_at_dst : t -> Base.Set.M(Vcs__.Path_in_repo).t

module Changed : sig ... end