Module Vcs.Non_raising

A functor to build non raising interfaces for Vcs based on a custom result type.

In addition to Or_error and Result, we provide this functor to create a Vcs interface based on a custom result type of your choice.

module type M = sig ... end

Interface used to build non raising interfaces to Vcs via Vcs.Non_raising.Make.

module type S = sig ... end

The interface exported by Vcs.

module Make (M : M) : S with type 'a t := 'a Vcs__.Vcs0.t and type 'a result := ('a, M.err) Base.Result.t

The interface exported by Vcs.