Module type Non_raising.M

Interface used to build non raising interfaces to Vcs via Vcs.Non_raising.Make.

type err

err must represent the type of errors in your monad.

The conversion functions you need to provide.

val map_error : Err.t -> err

val to_error : err -> Base.Error.t