Module Vcs.Refs

Capturing the information related to git refs.

In this file we represent the output of the git show-ref command. In particular we represent the location of all known local branches heads, remote branches heads and tags.

It is worth noting that the location of the remote branches is the state of the remote branches as they are currently known locally. To connect to the remote repository and retrieves information about the remote references, see Fetch_remote.

module Line : sig ... end

type t = Line.t Base.list

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val sexp_of_t : t -> Sexplib0.Sexp.t

Accessors

val tags : t -> Base.Set.M(Vcs__.Tag_name).t

val local_branches : t -> Branch_name.t Base.list

val remote_branches : t -> Remote_branch_name.t Base.list