Module Vcs.Tree

Building an in-memory representation of a git tree, for queries about history contents.

This data structure only contains the nodes, as well as the location of known branches (local and remotes) and tags.

The design is such that it is easy to add new nodes, and to compute the diff of what needs to be sent to a process holding such a value in memory to complete its view of a tree.

type t

val sexp_of_t : t -> Sexplib0.Sexp.t

val create : Base.unit -> t

create an empty tree that has no nodes.

val add_nodes : t -> log:Log.t -> Base.unit

add t log add to t all the nodes from the tree log. This is idempotent if t already knows all of the nodes.

module Node_kind : sig ... end

module Node : sig ... end

Refs

val refs : t -> Refs.t

List known refs.

val set_refs : t -> refs:Refs.t -> Base.unit

val set_ref : t -> rev:Rev.t -> ref_kind:Ref_kind.t -> Base.unit

val find_ref : t -> ref_kind:Ref_kind.t -> Node.t Base.option

Find a ref if it is present.

Revisions

val find_rev : t -> rev:Rev.t -> Node.t Base.option

Find the node pointed by the rev if any.

val mem_rev : t -> rev:Rev.t -> Base.bool

Tell if a revision points to a valid node of the tree.

Roots

val roots : t -> Node.t Base.list

Ancestors

val is_strict_ancestor : t -> ancestor:Node.t -> descendant:Node.t -> Base.bool

is_strict_ancestor t ~ancestor:a ~descendant:b returns true iif there exists a non empty path that leads from a to b. By definition, any node is not a strict ancestor of itself. .

val is_ancestor_or_equal : t -> ancestor:Node.t -> descendant:Node.t -> Base.bool

is_ancestor_or_equal t ~ancestor:a ~descendant:b returns true iif there a is a strict ancestor of b or if a is equal to b.

Tips

val tips : t -> Node.t Base.list

Return the list of nodes that do not have any children.

Log

val log : t -> Log.t

Subtree

module Subtree : sig ... end

val subtrees : t -> Subtree.t Base.list

val of_subtree : Subtree.t -> t

module Summary : sig ... end

val summary : t -> Summary.t

Print a summary for use in expect test and quick exploratory tests