Module Vcs.User_name

User name information as specified by the Git config user.name value.

Usually, this includes the full first name and last name, such as User_name.v "John Doe". This is not to be conflated with the User_handle, which are used as GitHub logins and other contexts.

type t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : t Base__Ppx_hash_lib.hash_fold

val hash : t -> Base__Ppx_hash_lib.Std.Hash.hash_value

val sexp_of_t : t -> Sexplib0.Sexp.t

val to_string : t -> Base.string

Given that t = string in the implementation, this function is just the identity.

val of_string : Base.string -> t Base.Or_error.t

of_string str returns Ok str if X.invariant str = true, and an error otherwise. This is meant to be used to validate untrusted entries.

val v : Base.string -> t

v str is a convenient wrapper for of_string str |> Or_error.ok_exn. This is typically handy for applying on trusted literals.